requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
package-name==0.1
parse==1.20.0
parse-type==0.6.2
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.16.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
